third_party/blink/web_tests/platform/generic/gamepad/gamepad-events-basic-expected.txt

Basic test for 'gamepadconnected' and 'gamepaddisconnected' events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Gamepad connected
PASS event.__proto__ is GamepadEvent.prototype
PASS event.__proto__.__proto__ is Event.prototype
PASS event.gamepad.id is "MockStick 3000"
PASS event.gamepad.buttons.length is 3
PASS event.gamepad.axes.length is 3
PASS event.gamepad.buttons[0].value is 1.0
PASS event.gamepad.buttons[0].pressed is true
PASS event.gamepad.buttons[1].value is 0.0
PASS event.gamepad.buttons[1].pressed is false
PASS event.gamepad.axes.length is 3
PASS event.gamepad.buttons[2].value is 0.333333
PASS event.gamepad.buttons[2].pressed is true
PASS event.gamepad.axes[0] is 0.5
PASS event.gamepad.axes[1] is -1.0
PASS event.gamepad.axes[2] is 0.333333
PASS axes[0] is 0.5
PASS buttons.hasOwnProperty('foo') is false
Gamepad disconnected
PASS event.__proto__ is GamepadEvent.prototype
PASS event.__proto__.__proto__ is Event.prototype
PASS event.gamepad.id is "MockStick 3000"
PASS event.gamepad.buttons.length is 3
PASS event.gamepad.axes.length is 3
PASS successfullyParsed is true

TEST COMPLETE
